HbsTranslationReference: Move into "intl" subpackage
